+++ src/main/java/com/ahead/school/dao/StudentDao.java
@@ -1,8 +1,7 @@
 package com.ahead.school.dao;
 
-
 import com.ahead.school.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentDao extends JpaRepository<Student,Long>{
+public interface StudentDao extends JpaRepository<Student, Integer> {
 }
